+++ tests/test_conversion.py
@@ -13,6 +13,7 @@ INT_EXCEPTED_NUMERAL_SYSTEMS = [
     NumeralSystem.ENGLISH_KEYCAP,
     NumeralSystem.ENGLISH_EMOJI,
     NumeralSystem.ENGLISH_FULL_STOP,
+    NumeralSystem.ENGLISH_COMMA,
     NumeralSystem.WANCHO,
     NumeralSystem.DIVES_AKURU,
     NumeralSystem.NYIAKENG_PUACHUE_HMONG]
@@ -34,6 +35,7 @@ CONVERSION_CASES = {
     NumeralSystem.ENGLISH_KEYCAP: "0️⃣1️⃣2️⃣3️⃣4️⃣5️⃣6️⃣7️⃣8️⃣9️⃣",
     NumeralSystem.ENGLISH_EMOJI: "0️1️2️3️4️5️6️7️8️9️",
     NumeralSystem.ENGLISH_FULL_STOP: "🄀⒈⒉⒊⒋⒌⒍⒎⒏⒐",
+    NumeralSystem.ENGLISH_COMMA: "🄁🄂🄃🄄🄅🄆🄇🄈🄉🄊",
     NumeralSystem.PERSIAN: "۰۱۲۳۴۵۶۷۸۹",
     NumeralSystem.HINDI: "०१२३४५६७८९",
     NumeralSystem.BENGALI: "০১২৩৪৫৬৭৮৯",