+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - English full stop mode
+- English comma mode
 ## [1.2] - 2025-12-11
 ### Added
 - English keycap mode